requirements.txt
Django==3.0
djangorestframework==3.11.0
psycopg2-binary>=2.8

requests==2.27.1
django-widget_tweaks